Playing around trying to get some sort of image cropper
Trying to get the beginnings of an image cropping functionality going.
Managed to get some JS going that lets the user draw a rectangle over
the image they  have uploaded. Now to try and get the coordinates of
that rectangle, crop the image accordingly and pass only the cropped
image to Cloud Vision. JS only works on a hardcoded local image right
now so need to make this work for whatever the user uploads. Need to try
and make it so we can do the cropping work on a local client side image
rather than uploading to server and downloading again because that doesn't
seem very efficient at all.