Add attempts at client side image handling
Trying to get client side image cropping working because saving to
the server before cropping is slow. Using the current form setup
doesn't really work because sending a POST request to /index triggers
the image save to server. Maybe change Upload Image button from form
submit to just a normal button. This way it won't trigger a POST request
and we can just link the JS cropping function to an onClick event on the
button. Then we link the POST request to some new upload button after the
user has done their cropping.

Also be careful with the form and CSRF tokens - when playing around with the
form HTML tag the CSRF token was appearing in the URL sometimes (I think when
removing the method="POST" and action="" elements from the form HTML tag)