Adds steps towards passing image dataURL to CloudVisionAPI
Makes first steps towards modifying app to pass a dataURL with the post cropped canvas content
to the Cloud Vision API rather than uploading a file. Uses Fetch API to do this. Currently Fetch
API returns a NetworkError. Suspect it may be something to do with CORS and running on a localhost
Flask server.